+++ doc/source/whatsnew/v0.20.2.txt
@@ -40,7 +40,7 @@ Bug Fixes
 - Bug in using ``pathlib.Path`` or ``py.path.local`` objects with io functions (:issue:`16291`)
 - Bug in ``DataFrame.update()`` with ``overwrite=False`` and ``NaN values`` (:issue:`15593`)
 - Passing an invalid engine to :func:`read_csv` now raises an informative
-  ValueError rather than UnboundLocalError. (:issue:`16511`)
+  ``ValueError`` rather than ``UnboundLocalError``. (:issue:`16511`)
